+++ tests/fix_loop.v
@@ -1,4 +1,4 @@
 From HB Require Import structures.
 
 HB.mixin Record M A := { x: nat }.
-#[primitive_class] HB.structure Definition S := { A of M A }.
+HB.structure Definition S := { A of M A }.